strata-render/backend/requirements.txt
# STRATA Backend — Python dependencies

# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Auth ──────────────────────────────────────────────────────────────────────
pyjwt==2.9.0
bcrypt==4.2.0

# ── AI — Google Gemini (free tier) ────────────────────────────────────────────
httpx==0.27.2

# ── Redis (optional — rate limiting) ─────────────────────────────────────────
redis[asyncio]==5.2.0

# ── Payments (optional) ───────────────────────────────────────────────────────
stripe==11.2.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
python-dotenv==1.0.1
